resource.c: update RT_Clean about close all tracked handles and clean handle list.

+++ src/memory.c
@@ -762,7 +762,7 @@ errno MT_Encrypt()
         num++;
     }
 
-    printf_s("num pages: %llu\n", pages->Len);
+    printf_s("num pages: %llu\n", (uint64)(pages->Len));
 
     // encrypt region and page list
     List* list = &tracker->Regions;
